backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "arghya-bandyopadhyay-30",email = "arghya.bandyopadhyay.official1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "setuptools (>=77.0.3,<78.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
